refactor: Split the profile screen into separate files
ProfileScreen.swift was the largest file in the app at 820 lines.
The state, the sections and the previews now each have their own
file, and the long state mapping in ProfileTab is broken up.
