+++ ios/Packages/Profile/Sources/Profile/ProfileScreen.swift
@@ -5,121 +5,15 @@ import SwiftUI
 import TvManiac
 
 public struct ProfileScreen: View {
-    public struct State: Equatable {
-        public let title: String
-        public let isLoading: Bool
-        public let userProfile: SwiftProfileInfo?
-        public let editButtonLabel: String
-        public let statsTitle: String
-        public let watchTimeLabel: String
-        public let monthsLabel: String
-        public let daysLabel: String
-        public let hoursLabel: String
-        public let episodesWatchedLabel: String
-        public let showsWatchedLabel: String
-        public let listsLabel: String
-        public let listsViewLabel: String
-        public let userListsTitle: String
-        public let viewAllLabel: String
-        public let retryLabel: String
-        public let userLists: SwiftSectionState<SwiftProfileList>
-        public let progressTitle: String
-        public let inProgressLabel: String
-        public let completedLabel: String
-        public let progressEmptyLabel: String
-        public let inProgress: SwiftSectionState<SwiftProfileShow>
-        public let completed: SwiftSectionState<SwiftProfileShow>
-        public let recentlyWatchedTitle: String
-        public let recentlyWatched: SwiftSectionState<SwiftProfileRecentShow>
-        public let favoritesTitle: String
-        public let favorites: SwiftSectionState<SwiftProfileShow>
-        public let unauthenticatedTitle: String
-        public let authTitle: String
-        public let authDescription: String
-        public let isAuthenticated: Bool
-        public let featureItems: [SwiftFeatureItem]
-        public let authProviders: [SwiftAuthProvider]
+    @Environment(\.appTheme) var appTheme
 
-        public init(
-            title: String,
-            isLoading: Bool,
-            userProfile: SwiftProfileInfo?,
-            editButtonLabel: String,
-            statsTitle: String,
-            watchTimeLabel: String,
-            monthsLabel: String,
-            daysLabel: String,
-            hoursLabel: String,
-            episodesWatchedLabel: String,
-            showsWatchedLabel: String,
-            listsLabel: String,
-            listsViewLabel: String,
-            userListsTitle: String = "",
-            viewAllLabel: String = "",
-            retryLabel: String = "",
-            userLists: SwiftSectionState<SwiftProfileList> = .empty,
-            progressTitle: String = "",
-            inProgressLabel: String = "",
-            completedLabel: String = "",
-            progressEmptyLabel: String = "",
-            inProgress: SwiftSectionState<SwiftProfileShow> = .empty,
-            completed: SwiftSectionState<SwiftProfileShow> = .empty,
-            recentlyWatchedTitle: String = "",
-            recentlyWatched: SwiftSectionState<SwiftProfileRecentShow> = .empty,
-            favoritesTitle: String = "",
-            favorites: SwiftSectionState<SwiftProfileShow> = .empty,
-            unauthenticatedTitle: String,
-            authTitle: String,
-            authDescription: String,
-            isAuthenticated: Bool,
-            featureItems: [SwiftFeatureItem],
-            authProviders: [SwiftAuthProvider] = []
-        ) {
-            self.title = title
-            self.isLoading = isLoading
-            self.userProfile = userProfile
-            self.editButtonLabel = editButtonLabel
-            self.statsTitle = statsTitle
-            self.watchTimeLabel = watchTimeLabel
-            self.monthsLabel = monthsLabel
-            self.daysLabel = daysLabel
-            self.hoursLabel = hoursLabel
-            self.episodesWatchedLabel = episodesWatchedLabel
-            self.showsWatchedLabel = showsWatchedLabel
-            self.listsLabel = listsLabel
-            self.listsViewLabel = listsViewLabel
-            self.userListsTitle = userListsTitle
-            self.viewAllLabel = viewAllLabel
-            self.retryLabel = retryLabel
-            self.userLists = userLists
-            self.progressTitle = progressTitle
-            self.inProgressLabel = inProgressLabel
-            self.completedLabel = completedLabel
-            self.progressEmptyLabel = progressEmptyLabel
-            self.inProgress = inProgress
-            self.completed = completed
-            self.recentlyWatchedTitle = recentlyWatchedTitle
-            self.recentlyWatched = recentlyWatched
-            self.favoritesTitle = favoritesTitle
-            self.favorites = favorites
-            self.unauthenticatedTitle = unauthenticatedTitle
-            self.authTitle = authTitle
-            self.authDescription = authDescription
-            self.isAuthenticated = isAuthenticated
-            self.featureItems = featureItems
-            self.authProviders = authProviders
-        }
-    }
-
-    @Environment(\.appTheme) private var appTheme
-
-    private let state: State
-    private let onSettingsClicked: () -> Void
-    private let onProviderSelected: (String) -> Void
-    private let onViewListsClicked: () -> Void
-    private let onRetryLists: () -> Void
-    private let onShowClicked: (Int64) -> Void
-    private let onRetryProgress: () -> Void
+    let state: State
+    let onSettingsClicked: () -> Void
+    let onProviderSelected: (String) -> Void
+    let onViewListsClicked: () -> Void
+    let onRetryLists: () -> Void
+    let onShowClicked: (Int64) -> Void
+    let onRetryProgress: () -> Void
 
     public init(
         state: State,
@@ -186,7 +80,7 @@ public struct ProfileScreen: View {
             VStack(spacing: 0) {
                 GeometryReader { proxy in
                     let scrollY = proxy.frame(in: .named("profileScroll")).minY
-                    let headerHeight = DimensionConstants.imageHeight + max(scrollY, 0)
+                    let headerHeight = ProfileDimensions.imageHeight + max(scrollY, 0)
 
                     headerContent(userProfile: userProfile, height: headerHeight)
                         .frame(width: proxy.size.width, height: headerHeight)
@@ -199,7 +93,7 @@ public struct ProfileScreen: View {
                             }
                         }
                 }
-                .frame(height: DimensionConstants.imageHeight)
+                .frame(height: ProfileDimensions.imageHeight)
 
                 VStack(spacing: 0) {
                     Spacer()
@@ -346,166 +240,12 @@ public struct ProfileScreen: View {
         }
     }
 
-    // MARK: - User Lists Section
-
-    @ViewBuilder
-    private var userListsSection: some View {
-        if case .empty = state.userLists {
-            EmptyView()
-        } else {
-            VStack(spacing: 0) {
-                Spacer().frame(height: appTheme.spacing.large)
-
-                CollapsibleSection(
-                    title: state.userListsTitle,
-                    showMore: userListsCount > DimensionConstants.maxInlineLists,
-                    onMoreClick: onViewListsClicked
-                ) {
-                    userListsBody
-                }
-            }
-        }
-    }
-
-    @ViewBuilder
-    private var userListsBody: some View {
-        switch state.userLists {
-        case .loading:
-            ScrollView(.horizontal, showsIndicators: false) {
-                HStack(spacing: appTheme.spacing.small) {
-                    ForEach(0 ..< 3, id: \.self) { _ in
-                        ShimmerView(cornerRadius: appTheme.shapes.large)
-                            .frame(width: 200, height: 133)
-                    }
-                }
-                .padding(.horizontal, appTheme.spacing.medium)
-            }
-        case let .error(message):
-            InlineSectionError(
-                message: message,
-                retryLabel: state.retryLabel,
-                onRetry: onRetryLists
-            )
-        case let .content(lists):
-            ScrollView(.horizontal, showsIndicators: false) {
-                HStack(spacing: appTheme.spacing.small) {
-                    ForEach(lists) { list in
-                        ListCollageCard(list: list, onClick: {})
-                    }
-                }
-                .padding(.horizontal, appTheme.spacing.medium)
-            }
-        case .empty:
-            EmptyView()
-        }
-    }
-
-    private var userListsCount: Int {
-        if case let .content(lists) = state.userLists {
-            return lists.count
-        }
-        return 0
-    }
-
-    // MARK: - Progress Section
-
-    @ViewBuilder
-    private var progressSection: some View {
-        if case .empty = state.inProgress, case .empty = state.completed {
-            EmptyView()
-        } else {
-            VStack(spacing: 0) {
-                Spacer().frame(height: appTheme.spacing.large)
-
-                ProgressSectionView(
-                    inProgress: state.inProgress,
-                    completed: state.completed,
-                    title: state.progressTitle,
-                    inProgressLabel: state.inProgressLabel,
-                    completedLabel: state.completedLabel,
-                    emptyLabel: state.progressEmptyLabel,
-                    retryLabel: state.retryLabel,
-                    onShowClick: onShowClicked,
-                    onRetry: onRetryProgress
-                )
-            }
-        }
-    }
-
-    // MARK: - Recently Watched Section
-
-    @ViewBuilder
-    private var recentlyWatchedSection: some View {
-        if case .empty = state.recentlyWatched {
-            EmptyView()
-        } else {
-            VStack(spacing: 0) {
-                Spacer().frame(height: appTheme.spacing.large)
-
-                RecentlyWatchedSectionView(
-                    recentlyWatched: state.recentlyWatched,
-                    title: state.recentlyWatchedTitle,
-                    retryLabel: state.retryLabel,
-                    onShowClick: onShowClicked,
-                    onRetry: onRetryProgress
-                )
-            }
-        }
-    }
-
-    // MARK: - Favorites Section
-
-    @ViewBuilder
-    private var favoritesSection: some View {
-        if case .empty = state.favorites {
-            EmptyView()
-        } else {
-            VStack(spacing: 0) {
-                Spacer().frame(height: appTheme.spacing.large)
-
-                FavoritesSectionView(
-                    favorites: state.favorites,
-                    title: state.favoritesTitle,
-                    retryLabel: state.retryLabel,
-                    onShowClick: onShowClicked,
-                    onRetry: onRetryProgress
-                )
-            }
-        }
-    }
-
-    private func bigCount(_ value: Int) -> some View {
-        StatValueText(count: value)
-    }
-
-    private func bigLabel(_ text: String) -> some View {
-        Text(text)
-            .textStyle(appTheme.typography.headlineLarge)
-            .foregroundStyle(.appOnSurface)
-            .shadow(color: .black.opacity(0.3), radius: 4, x: 0, y: 2)
-            .lineLimit(1)
-            .minimumScaleFactor(0.5)
-    }
-
-    private func watchTimeSegment(value: Int32, unit: String) -> some View {
-        HStack(alignment: .firstTextBaseline, spacing: appTheme.spacing.xxxSmall) {
-            StatValueText(count: Int(value))
-                .lineLimit(1)
-                .minimumScaleFactor(0.5)
-
-            Text(unit)
-                .textStyle(appTheme.typography.bodySmall)
-                .foregroundStyle(appTheme.colors.onSurfaceVariant)
-                .lineLimit(1)
-        }
-    }
-
     // MARK: - Loading Skeleton
 
     private var profileSkeleton: some View {
         VStack(spacing: 0) {
             ShimmerView(cornerRadius: 0)
-                .frame(height: DimensionConstants.imageHeight)
+                .frame(height: ProfileDimensions.imageHeight)
 
             VStack(spacing: appTheme.spacing.small) {
                 HStack {
@@ -604,217 +344,4 @@ public struct ProfileScreen: View {
             Spacer()
         }
     }
-}
-
-private struct StatValueText: View {
-    @Environment(\.appTheme) private var appTheme
-
-    let count: Int
-
-    var body: some View {
-        AnimatedCountText(count: count)
-            .textStyle(appTheme.typography.headlineLarge)
-            .foregroundStyle(.appOnSurface)
-            .shadow(color: .black.opacity(0.3), radius: 4, x: 0, y: 2)
-    }
-}
-
-private enum DimensionConstants {
-    static let imageHeight: CGFloat = 315
-    static let maxInlineLists: Int = 4
-}
-
-private struct ProfileScrollOffsetKey: PreferenceKey {
-    static var defaultValue: CGFloat = 0
-
-    static func reduce(value: inout CGFloat, nextValue: () -> CGFloat) {
-        value = nextValue()
-    }
-}
-
-private let previewFeatureItems: [SwiftFeatureItem] = [
-    SwiftFeatureItem(
-        id: "discover",
-        iconName: "magnifyingglass",
-        title: "Discover",
-        description: "Browse over one million movies and TV shows."
-    ),
-    SwiftFeatureItem(
-        id: "track",
-        iconName: "tv",
-        title: "Track",
-        description: "Check-in, mark watched and manage your watch history."
-    ),
-    SwiftFeatureItem(
-        id: "manage",
-        iconName: "rectangle.stack",
-        title: "Watchlist",
-        description: "Create custom lists, personalize just the way you like."
-    ),
-    SwiftFeatureItem(
-        id: "more",
-        iconName: "sparkles",
-        title: "More",
-        description: "More features coming soon."
-    ),
-]
-
-private let previewShows: [SwiftProfileShow] = [
-    SwiftProfileShow(id: 1, title: "Breaking Bad", posterUrl: nil),
-    SwiftProfileShow(id: 2, title: "Game of Thrones", posterUrl: nil),
-    SwiftProfileShow(id: 3, title: "Stranger Things", posterUrl: nil),
-]
-
-private let previewRecentShows: [SwiftProfileRecentShow] = [
-    SwiftProfileRecentShow(showId: 1, title: "Breaking Bad", posterUrl: nil, episodeLabel: "S5E14"),
-    SwiftProfileRecentShow(showId: 2, title: "Game of Thrones", posterUrl: nil, episodeLabel: "S8E3"),
-]
-
-#Preview("Authenticated – Stats Present") {
-    ProfileScreen(
-        state: ProfileScreen.State(
-            title: "Profile",
-            isLoading: false,
-            userProfile: SwiftProfileInfo(
-                username: "tvmaniac_user",
-                fullName: "John Doe",
-                avatarUrl: nil,
-                backgroundUrl: nil,
-                stats: SwiftProfileStats(
-                    showsWatched: "87",
-                    episodesWatched: "1,250",
-                    months: 2,
-                    days: 15,
-                    hours: 8,
-                    listCount: 12
-                )
-            ),
-            editButtonLabel: "Edit Profile",
-            statsTitle: "Stats",
-            watchTimeLabel: "Watch Time",
-            monthsLabel: "M",
-            daysLabel: "D",
-            hoursLabel: "H",
-            episodesWatchedLabel: "Episodes Watched",
-            showsWatchedLabel: "Shows Watched",
-            listsLabel: "Lists",
-            listsViewLabel: "View",
-            userListsTitle: "Your Lists",
-            viewAllLabel: "More",
-            retryLabel: "Retry",
-            inProgress: .content(previewShows),
-            completed: .content(previewShows),
-            recentlyWatchedTitle: "Recently Watched",
-            recentlyWatched: .content(previewRecentShows),
-            favoritesTitle: "Favorites",
-            favorites: .content(previewShows),
-            unauthenticatedTitle: "Discover.\nTrack.\nWatchlist.\n& More ...",
-            authTitle: "Connect & Sync Your Content",
-            authDescription: "Save your progress, discover new titles, and sync across all devices.",
-            isAuthenticated: true,
-            featureItems: previewFeatureItems
-        ),
-        onSettingsClicked: {},
-        onProviderSelected: { _ in }
-    )
-}
-
-#Preview("Authenticated – Stats Hidden (Simkl)") {
-    ProfileScreen(
-        state: ProfileScreen.State(
-            title: "Profile",
-            isLoading: false,
-            userProfile: SwiftProfileInfo(
-                username: "simkl_user",
-                fullName: "Jane Doe",
-                avatarUrl: nil,
-                backgroundUrl: nil,
-                stats: nil
-            ),
-            editButtonLabel: "Edit Profile",
-            statsTitle: "Stats",
-            watchTimeLabel: "Watch Time",
-            monthsLabel: "M",
-            daysLabel: "D",
-            hoursLabel: "H",
-            episodesWatchedLabel: "Episodes Watched",
-            showsWatchedLabel: "Shows Watched",
-            listsLabel: "Lists",
-            listsViewLabel: "View",
-            userListsTitle: "Your Lists",
-            viewAllLabel: "More",
-            retryLabel: "Retry",
-            inProgress: .content(previewShows),
-            completed: .content(previewShows),
-            recentlyWatchedTitle: "Recently Watched",
-            recentlyWatched: .content(previewRecentShows),
-            favoritesTitle: "Favorites",
-            favorites: .content(previewShows),
-            unauthenticatedTitle: "Discover.\nTrack.\nWatchlist.\n& More ...",
-            authTitle: "Connect & Sync Your Content",
-            authDescription: "Save your progress, discover new titles, and sync across all devices.",
-            isAuthenticated: true,
-            featureItems: previewFeatureItems
-        ),
-        onSettingsClicked: {},
-        onProviderSelected: { _ in }
-    )
-}
-
-#Preview("Loading") {
-    ProfileScreen(
-        state: ProfileScreen.State(
-            title: "Profile",
-            isLoading: true,
-            userProfile: nil,
-            editButtonLabel: "Edit Profile",
-            statsTitle: "Stats",
-            watchTimeLabel: "Watch Time",
-            monthsLabel: "M",
-            daysLabel: "D",
-            hoursLabel: "H",
-            episodesWatchedLabel: "Episodes Watched",
-            showsWatchedLabel: "Shows Watched",
-            listsLabel: "Lists",
-            listsViewLabel: "View",
-            unauthenticatedTitle: "Discover.\nTrack.\nWatchlist.\n& More ...",
-            authTitle: "Connect & Sync Your Content",
-            authDescription: "Save your progress, discover new titles, and sync across all devices.",
-            isAuthenticated: false,
-            featureItems: previewFeatureItems
-        ),
-        onSettingsClicked: {},
-        onProviderSelected: { _ in }
-    )
-}
-
-#Preview("Unauthenticated") {
-    ProfileScreen(
-        state: ProfileScreen.State(
-            title: "Profile",
-            isLoading: false,
-            userProfile: nil,
-            editButtonLabel: "Edit Profile",
-            statsTitle: "Stats",
-            watchTimeLabel: "Watch Time",
-            monthsLabel: "M",
-            daysLabel: "D",
-            hoursLabel: "H",
-            episodesWatchedLabel: "Episodes Watched",
-            showsWatchedLabel: "Shows Watched",
-            listsLabel: "Lists",
-            listsViewLabel: "View",
-            unauthenticatedTitle: "Discover.\nTrack.\nWatchlist.\n& More ...",
-            authTitle: "Connect & Sync Your Content",
-            authDescription: "Save your progress, discover new titles, and sync across all devices.",
-            isAuthenticated: false,
-            featureItems: previewFeatureItems,
-            authProviders: [
-                SwiftAuthProvider(id: "TRAKT", label: "Continue with Trakt", logoName: "TraktMono"),
-                SwiftAuthProvider(id: "SIMKL", label: "Continue with Simkl", logoName: "SimklMono"),
-            ]
-        ),
-        onSettingsClicked: {},
-        onProviderSelected: { _ in }
-    )
 }
